refactor(core): fix core module imports
