+++ projects/dsp-ui/src/lib/action/components/login-form/login-form.component.spec.ts
@@ -1,6 +1,6 @@
 import { Component, OnInit, ViewChild } from '@angular/core';
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
-import { FormBuilder, FormsModule, ReactiveFormsModule} from '@angular/forms';
+import { FormBuilder, ReactiveFormsModule } from '@angular/forms';
 import { MatInputModule } from '@angular/material/input';
 import { MatSnackBarModule } from '@angular/material/snack-bar';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
@@ -14,7 +14,8 @@ import {
 } from '@dasch-swiss/dsp-js';
 import { of } from 'rxjs';
 import { AjaxResponse } from 'rxjs/ajax';
-import { DspApiConnectionToken, Session, SessionService } from '../../../core';
+import { DspApiConnectionToken } from '../../../core/core.module';
+import { Session, SessionService } from '../../../core/session.service';
 import { LoginFormComponent } from './login-form.component';
 
 /**
