+++ projects/dsp-ui/src/lib/search/advanced-search/select-property/specify-property-value/search-list-value/search-list-value.component.ts
@@ -1,6 +1,6 @@
 import { Component, Inject, Input, OnDestroy, OnInit, ViewChild } from '@angular/core';
-import { IRI, PropertyValue, Value } from '../operator';
 import { FormBuilder, FormGroup, Validators } from '@angular/forms';
+import { MatMenuTrigger } from '@angular/material/menu';
 import {
     ApiResponseError,
     Constants,
@@ -8,9 +8,9 @@ import {
     ListNodeV2,
     ResourcePropertyDefinition
 } from '@dasch-swiss/dsp-js';
-import { MatMenuTrigger } from '@angular/material/menu';
-import { DspApiConnectionToken } from '../../../../../core';
-import { NotificationService } from '../../../../../action';
+import { NotificationService } from '../../../../../action/services/notification.service';
+import { DspApiConnectionToken } from '../../../../../core/core.module';
+import { IRI, PropertyValue, Value } from '../operator';
 
 // https://stackoverflow.com/questions/45661010/dynamic-nested-reactive-form-expressionchangedafterithasbeencheckederror
 const resolvedPromise = Promise.resolve(null);
